Fix Linux/OS-X compile errors (issue #751).
git-svn-id: https://chromiumembedded.googlecode.com/svn/branches/1180@896 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ cef1/include/internal/cef_types_wrappers.h
@@ -507,8 +507,8 @@ class CefTime : public CefStructBase<CefTimeTraits> {
   }
 
   // Return the delta between this object and |other| in milliseconds.
-  int64 Delta(const CefTime& other) {
-    int64 delta = 0;
+  long long Delta(const CefTime& other) {
+    long long delta = 0;
     cef_time_delta(this, &other, &delta);
     return delta;
   }
